Progress in proper database setup

+++ app/src/main/java/com/mycarni_garden/data/model/ComponentSubstrateCrossRef.java
@@ -20,7 +20,7 @@ public class ComponentSubstrateCrossRef {
     public long component_id;
     public long substrate_id;
 
-    public ComponentSubstrateCrossRef(long component_id, long substrate_id) {
+    public ComponentSubstrateCrossRef(long component_id, int substrate_id) {
         this.component_id = component_id;
         this.substrate_id = substrate_id;
     }
